Refactoring function to create new order messages out of bot.js
